+++ CHANGELOG.md
@@ -7,6 +7,25 @@ If it saves you or your team time, please consider [sponsoring its development](
 The namespace remains `Kreait\Firebase` and the package name remains `kreait/firebase-php`.
 Please update your remote URL if you have forked or cloned the repository.
 
+## Unreleased
+
+### Messaging
+
+The FCM API [deprecated](https://firebase.google.com/docs/reference/fcm/rest/v1/projects.messages) the
+`token` field of a message in favor of the new `fid` field, which targets a message at a
+[Firebase Installation ID](https://firebase.google.com/docs/cloud-messaging/android/get-started#access-firebase-installation-id).
+The `token` field remains fully supported until Google decommissions it, and during the transition
+period it also accepts Firebase Installation IDs - no immediate action is required.
+
+* Added `Kreait\Firebase\Messaging\CloudMessage::withFid()` and support for the `fid` key in
+  `CloudMessage::fromArray()`. Like the other targets, a FID replaces a previously set target.
+* Added the `Kreait\Firebase\Messaging\FirebaseInstallationId` and
+  `Kreait\Firebase\Messaging\FirebaseInstallationIds` value objects.
+* Added `Kreait\Firebase\Messaging\MessageTarget::FID`.
+* `Messaging::sendMulticast()` now also accepts `FirebaseInstallationId(s)` instances. Strings and arrays
+  of strings continue to be treated as registration tokens.
+* Added `MulticastSendReport::validFids()` and `MulticastSendReport::unknownFids()`.
+
 ## 8.5.0 - 2026-09-13
 
 Added support for providing a PSR-14 event dispatcher to the factory.